Add files via upload
Block is a superclass, and EditableTextBlock and NonEditableTextBlock are it's two children.
Block needs to be modified and throw an exception in the constructor to only accept legal ASCII values that are the 26 uppercase letters in the English alphabet.
Block also needs to be modified to throw an exception to accept a legal int within the range of possible RGB color values (it might be easier to use hexadecimal for this).
Grid is a work in progress: it doesn't compile because of problems with the constructor. I need to find a better way to recursively add all NonEditableTextBlock's to an ArrayList.